+++ CHANGELOG.md
@@ -18,6 +18,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v5.12.2 - 2022-04-06
+
+### Fixed
+
+- 転送アクションのagent.urlとrecipient.urlがnullになる状態を解消
+
 ## v5.12.1 - 2022-04-06
 
 ### Changed